Added serde support.
